Quiet down the compiler.

+++ src/widgets/file_menu.cpp
@@ -78,7 +78,7 @@ void file_menu::display_current_files() {
 		// Make sure that all lines fit.
 		// Guess the width of the scrollbar to be 30 since it is not accessible from here.
 		// -25 to compensate for the picture column.
-		while (font::line_width(*it, menu_font_size) > width() - 30 - 25) {
+		while ((unsigned int)font::line_width(*it, menu_font_size) > width() - 30 - 25) {
 			(*it).resize((*it).size() - 1);
 		}
 	}
